+++ src/greatday/_runners.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from typing import Final, List
 
-import clack
 from clack.types import ClackRunner
 from logrus import Logger
+import metaman
 
 from ._common import CTX_INBOX, drop_words
 from ._config import AddConfig, ListConfig, TUIConfig
@@ -17,7 +17,7 @@ from ._tui import start_textual_app
 
 
 ALL_RUNNERS: List[ClackRunner] = []
-runner = clack.register_runner_factory(ALL_RUNNERS)
+runner = metaman.register_function_factory(ALL_RUNNERS)
 
 logger = Logger(__name__)
 
@@ -29,7 +29,7 @@ def run_add(cfg: AddConfig) -> int:
     """Runner for the 'add' subcommand."""
     log = logger.bind_fargs(locals())
 
-    repo = SQLRepo(cfg.database_url)
+    repo = SQLRepo(cfg.database_url, verbose=cfg.verbose)
     todo = GreatTodo.from_line(cfg.todo_line).unwrap()
 
     x_found = False
@@ -57,7 +57,7 @@ def run_add(cfg: AddConfig) -> int:
 @runner
 def run_list(cfg: ListConfig) -> int:
     """Runner for the 'list' subcommand."""
-    repo = SQLRepo(cfg.database_url)
+    repo = SQLRepo(cfg.database_url, verbose=cfg.verbose)
 
     query: str
     if cfg.query is None:
@@ -77,5 +77,5 @@ def run_list(cfg: ListConfig) -> int:
 @runner
 def run_tui(cfg: TUIConfig) -> int:
     """Runer for the 'tui' subcommand."""
-    start_textual_app(cfg.database_url, cfg.data_dir)
+    start_textual_app(cfg.database_url, verbose=cfg.verbose)
     return 0