+++ tests/test_add.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
-from .conftest import MainType
+from . import common
 
 
-def test_add(main: MainType) -> None:
+def test_add(main: common.MainType) -> None:
     """Tests the 'add' subcommand."""
     assert main("add", "foo") == 0